packing value item

+++ app/src/main/java/com/partsystem/partvisitapp/feature/create_order/model/FinalFactorRequest.kt
@@ -41,7 +41,7 @@ data class FinalFactorDetailDto(
     val price: Int,
     val description: String?,
     val packingId: Int?,
-    val packingValue: Int,
+    val packingValue: Double,
     val vat: Int,
     val productSerial: Int,
     val isGift: Int,
